Fix a crash when calling SHOW SLAVE STATUS while using a detached session
Summary: The code in rpl_slave.cc that was run when SHOW SLAVE STATUS was executed accessed the thd's packet buffer directly.  Since this was a detached session we want it to use the information in the connection thd's packet buffer.  This was available via the protocol, so I switched over to access the packet (and net) via the protocol object.  Also added a test that ran a bunch of SHOW commands to make sure none of the other ones had a similar problem.

Reviewed By: jrahman

Differential Revision: D7783389

fbshipit-source-id: 88df90a