Fix reset of THD query when trx metadata is enabled
Summary:
When a Rows_query_log_event contains trx metadata the real query is
present just after the metadata comment. Due to this the THD is assigned the
query after going past the metadata in Rows_query_log_event::do_apply_event()
but the destructor was resetting the query by comparing the THD query to the
beginning of the rows query. This was causing a read-after-free error.

Reviewed By: yashtc

Differential Revision: D7899271

fbshipit-source-id: 8374804