Check net before dereference
Summary:
In the calling function, len is initialized to 0 and never
changed if the net->vio is null. It then hits this branch and is
guaranteed to seg fault. This was less likely to happen in 5.6.35
since the dereference was inside the DBUG_PRINT statement, which is
usually optimized out

There seem to be other observable changes in the API here, but they're
all in UB cases

Reviewed By: lth

Differential Revision: D7782985

fbshipit-source-id: 88f04ca20c9