Report SSL timeout errors correctly
Summary:
Currently, when connect timeout occurs in SSL, these get incorrectly reported as :
  ERROR 2026 (HY000): SSL connection error: error:00000002:invalid library (0):OPENSSL_internal:system library

The reason is because when a timeout happens, ssl_errno_holder is still left with a WANT_READ error number from before the wait. This is then read by the caller of sslconnect and interpreted incorrectly. The fix is to have new error codes for timeouts and check for those, and ignore the ssl error when timeout is detected.

Reference patch: https://github.com/facebook/mysql-5.6/commit/1da2ce0

Reviewed By: abal147

Differential Revision: D7976483

fbshipit-source-id: ec5e2d30a20