Maintain a whitelist of semi-sync slaves which can ACK transactions
Summary:
A whitelist of semi-sync slaves can be maintained using a new sys-var
`rpl_semi_sync_master_whitelist`. The value of this variable is a comma
separated list of UUIDs. ACK from any slave which is not in the whitelist will
be ignored. The whitelist is persisted in file
`semi_sync_master_whitelist.info` to survive server restarts.

Reviewed By: Tema

Differential Revision: D7715511

fbshipit-source-id: 9be4edd