Fix "Error 1 during COMMIT" error
Summary:
The query_id for the detached session's THD was never getting set correctly.  This is the piece of information that is included in the RocksDB transaction to make it unique.  Updated the code to copy it from the connection's THD.

The new test failed every time with the "Error 1 during COMMIT" error and succeeds with the fix in place.

Reviewed By: AjanthanAsogamoorthy

Differential Revision: D8013383

fbshipit-source-id: bb019cc