Bug #27399620: BINLOG AND ENGINE BECOME INCONSISTENT WHEN BINLOG CACHE FILE GETS OUT OF SPACE
Summary:
Patch is from Oracle to resolve https://bugs.mysql.com/bug.php?id=89272

Problem:
--------
If the transaction size is bigger than the binlog_cache_size,
server uses a file in 'tmpdir' for flushing the data. If the
tmpdir partition is full, flush will fail with 'No Space'
error.

In such scenarios,
1. Flush errors were not handled properly in the
   gtid_before_write_cache function.

2. After preparing the transaction, client disconnection
   and rollback were hitting DBUG_ASSSERT(is_binlog_empty())
   in binlog_cache_data::reset() function.

Description:
------------

1. In the gtid_before_write_cache function, while flushing GTID
   to the cache,

   a. write position of the cache is set to position 0.
   b. Gtid_log_event is flushed to the cache.
   c. write position is reset to the old position.

In step 'a', reset_write_pos is invoked to set the write position
to 0. If the cache has swapped to a file, it reinitializes it by
calling reinit_io_cache() function.

While reinitializing the cache, it tries to flush the cache to
the file(file in tmpdir). Since the tmpdir partition is full, flush
fails with 'No Space' error and returns from that function.

Since reset_write_pos() or gtid_before_write_cache functions do
not to check if there was a failure in resetting the position,
the functions assume that the reset has happened and continues
without returning immediately and hits an assert
DBUG_ASSERT(!do_checksum || remains == 0) in
MYSQL_BIN_LOG::do_write_cache() function.

In release builds, though asserts do not happen, on flush
errors, we expected an entry in the error log saying that there
was an error while writing to binary log. But no such entry was
there in the error log.

2. On client disconnection/rollback, the server rolls back the
transaction. While doing so, it calls the
binlog_cache_data::reset() which in turn calls the truncate()
function wherein it truncates the binlog cache.

truncate() function tries to reinitialize the cache and
reinitialization fails(my_b_flush_io() fails). As a result,
binlog_cache is not cleared and hits the
DBUG_ASSERT(is_binlog_empty()).

Fix:
----
1. Flush errors have been handled in the
   gtid_before_write_cache() function.

   Now, the server will decide what action to take if this
   situation happens.

   If binlog_error_action == ABORT_SERVER, then it will abort the
   server after informing the client with
   'ER_BINLOGGING_IMPOSSIBLE' error.

   If binlog_error_action == 'IGNORE_ERROR', then it will ignore
   the error and disable the binlog  further until server is
   restarted again. The same will be mentioned in the error log
   file.

2. Two new functions have been added (set_flush_error and
   has_flush_error) to set and check the flush errors.
   On rollback/client disconnection, the check for
   flush_error is done and the cache is cleared in case of
   flush_errors.

Reviewed By: lth

Differential Revision: D8499181

Pulled By: hermanlee

fbshipit-source-id: 7be1c85