Fix multi-query when using detached sessions
Summary:
The server_status flag is duplicated in the connection THD and session THD.  The code was updating the session THD version (when using detached sessions), but when building the output packet, the protocol referenced the connection THD's version.  Since this didn't have any of the updates, the flag to indicate that more results were coming was not set.  Update the code to pass in the correct THD to Protocol::end_statement() and Protocol::send_output_parameters().

This problem would have affected all server status flags - not just more result sets

Reviewed By: jrahman

Differential Revision: D8195085

fbshipit-source-id: f59ba42