Fix a bug in wrong nonnull attribute
Summary:
Declaration of function `thd_requested_durability` had `nonnull` attribute, which means that its arguments are supposed to be nonnull values. However, the attribute is actually wrong and crashed mysqld during the `mysqltest.sh` run. I inserted printf before the callsite and confirmed that the parameter is `(nil)`: https://fburl.com/tsxalteo.

P59654596 is the bootstrap log before the fix (which crashes), and P59654696 is the log from the same run after the fix.

Reviewed By: hermanlee

Differential Revision: D8213263

fbshipit-source-id: c3cd748