Issue #834: Server crash in print_keydup_error / key_unpack or unexpected ER_DUP_KEY
The issue is caused by a combination of two factors:

1. Rdb_ddl_manager::rename() loses the value of m_hidden_pk_val. new
object used to get 0, which means "not loaded from the db yet".

2. ha_rocksdb::load_hidden_pk_value() uses current transaction (and its
snapshot) when loading hidden PK value from disk. This may cause it to
load an out-of-date value.

This patch fixes #1, which is sufficient to fix the issue.
(This is attempt 2, with review input addressed)