Fix MyRocks SK updates when master is missing columns
Summary:
When master has missing columns in the middle of the table and COMPLETE binlog
format is used SKs were not updated on properly on the slave. This was
happening because a temporary code was resetting the tables's write set after
it was set correctly using column names and MyRocks uses the write set to
compute SKs that need to be updated.

Removed the temporary code to fix the issue.

Differential Revision: D8457633 (https://github.com/facebook/mysql-5.6/commit/d29492d28cf5b434496e06a2505a2c99f6677f43)

fbshipit-source-id: 7bc497d4cab