Fix ASAN failures with use after scope error
Summary:
As clang ASAN is more strict that gcc ASAN, it detects additional bugs that gcc ASAN couldn't find.

* storage/innobase/handler/ha_innodb.cc

It is obvious that `nz2` is used outside of its scope when goto branch is taken.

* sql/sql_update.cc

In line 688 of sql_update.cc, `tempfile` is copied to `select->file`, but it's not a deep copy. In line 705 when `init_read_record` is called, the function may call `reinit_io_cache` with `select->file` as a parameter (line 218 of sql/records.cc), and `reinit_io_cache` dereferences `(IO_CACHE*)info->current_pos` which results use after scope.

Reviewed By: hermanlee

Differential Revision: D8626182

fbshipit-source-id: 81dfd17