Enable LLVM sanitizers
Summary:
This diff is to enable sanitizers (asan, ubsan, tsan) with LLVM. Inlined comments describe the reason for each change.

To run sanitizer, we need to update the commit hash for MyRocks, as we need to incorporate D8668941. But I didn't update the commit hash for RocksDB in this diff as it may have unexpected side effect.

Reviewed By: hermanlee

Differential Revision: D8678300

fbshipit-source-id: 92f2fae