Removed including lf_hash.c from connection_delay
Summary:
The connection_control tests were failing in asan due to symbol
collisions and differences in whether lf_hash symbols were linked in.
For builds with performance_schema, lf_hash functions were used by
mysqld, so they were pulled in. For builds without performance_schema,
they were not being pulled in. The connection_control plugin references
the lf library, but may fail to find them in non-performance builds.

Within mysqld.cc, reference one of the variables from lf_hash.c, which
then pulls in those definitions for non-performance schema builds.

An alternative way to solve this would be to provide the linker with
lf_hash_init as an undefined symbol, but this seems easier to do.

Reviewed By: jkedgar

Differential Revision: D8905206

fbshipit-source-id: a6f77cd40b2