Remove unflushed events warning when trx metadata is enabled on SBR
Summary:
mysqlbinlog expects rows_query event only in RBR. It prints an
unflushed event warning when it parses a transaction which has rows_query but
not a rows events with STMT_END_F flag set. Since there is no direct way to
tell the binlog format in mysqlbinlog, we'll just forgo the warning for
rows_event which contain metadata.

Squash with D9003508

Reviewed By: yashtc

Differential Revision: D9074409

fbshipit-source-id: 363fc8ef651