Check MySQL version when printing rocksdb_bulk_load session variables in mysqldump
Summary:
As we prepare to migrate to MySQL 8.0, update the check for RockDB support to ensure that dumps from 5.6 can be loaded into 8.0.

Adding this patch to 8.0 will also make sure that dumps from 8.0 will
enable rocksdb_bulk_load in MyRocks.

The controller you requested could not be found.: rocksdb

Reviewed By: lth

Differential Revision: D9136323

fbshipit-source-id: 0ab7d8900c5