Extending mysqlbinlog to filter empty transactions
Summary:
This diff adds an additional command line option --skip-empty-trans to be used in the case when both --database and --skip-gtids options are specified. Further logics for skipping the empty transactions of unrelated databases are added and enabled only when --database is selected and --skip-gtids are enabled. If either --database or --skip-gtids is not specified, mysqlbinlog will output an error message.

The fix covers both code paths of the statement based replication as well as the row based replication. The fix essentially uses a database string to keep track of the operating database of the current transaction. Since XID event does not have the corresponding database information, the database tracking string is used for filtering out the
COMMIT XID event in both statement and row based replication paths.

The fix also ensures that the database will not change in the middle of a transaction and will issue errors if it does. It also ensures that the database string is always cleared after a transaction and reassigned the new value when a new transaction starts.

Please note that the flag print_event_info->skipped_event_in_transaction (was set in the code path I touched in the diff) is to tell the next GTID event that there are skipped event previously so that the GTID event can automatically adds a COMMIT. Since --skip-empty-trans can only used with --skip-gtids, the GTID event will be skipped. In other words, the GTID code path checking print_event_info->skipped_event_in_transaction will not be touched in our case. So the flag of print_event_info->skipped_event_in_transaction will not affect the changes in this diff.

Also note that the only other place that checks print_event_info->skipped_event_in_transaction is at the very end after printing all the events. print_event_info->skipped_event_in_transaction is checked to print an automatically generated COMMIT message if the previous transaction has skipped events but the transaction itself has not hit a COMMIT event (print_event_info->skipped_event_in_transaction is always set to be false at the COMMIT event). Since --skip-empty-trans only supports a single database within a transaction, the entire transaction is either skipped or not skipped. Therefore, we do not need an automatically generated COMMIT message at the end if a skipped transaction has been interrupted in the middle.

Reference patch: https://github.com/facebook/mysql-5.6/commit/2bc7ebc

Reviewed By: lth

Differential Revision: D9310534

fbshipit-source-id: 3557b9ba44a