Supporting O_DIRECT reads in xbstream
Summary:
This diff adds an option -d to support reading
src files with O_DIRECT in xbstream. This is useful when copying
src data files by xbstream and if src instances are running with
O_DIRECT. MyRocks (myrocks_hotbackup) falls into that type so
it can be useful in MyRocks.

Reviewed By: anirbanr-fb

Differential Revision: D9502939

fbshipit-source-id: 4ff4f20ee14