Fixed valgrind issues
Summary:
When testing if a invalid semi-sync acker's dump thread is killed when
an ACK is recieved from it, we were not making sure that the ACK
received is from the invalid slave. So, sometimes in lengthy valgrind
runs we would send the transaction to the invalid slave after the valid
slave ACKed the transaction. Since, any transaction sent to semi-sync
slaves after an ACK will not have the magic header set, the invalid
slave will never send an ACK. So the dump thread never dies and we
timeout in our test while waiting for it to die.

Fixed the test by stopping the valid slave so we receive ACK from invalid one.

Reviewed By: Tema

Differential Revision: D9551467

fbshipit-source-id: 413497d7522