Stablize main.com_rpc
Summary:
Fixing the remaining places where we check processlist without first waiting for the query to start executing, or waiting for query to be killed.

There is also a race condition between KILL and session attach that is causing test failures. This diff does not fix that race condition, but modifies the test to avoid that race instead.

Reviewed By: abal147

Differential Revision: D10438857

fbshipit-source-id: 40120d7fd53