Audit session_tracker information
Summary:
Add the concept of auditing the session tracker. At the moment, we will only audit the response attributes, but the capability is there to audit any of them.

Each session tracker implements an interface which allows it to pass arbitrary key/value pairs to the audit plugin. The audit plugin will receive this as a char** array, where each key is followed immediately by it's value. query attributes are similarly passed to the plugin as a char** array of [key1, value1, ... ,keyN, ValueN].

Reviewed By: lth

Differential Revision: D10501605

fbshipit-source-id: 2c4bda690e6