Fix flakiness in binlog.binlog_rows_query_with_filter
Summary:
Disable warnings when calling purge binary logs because the active log
will not be purged. Depending on the amount of time between flush logs
and purge binary logs, the server may have created a new log file and
generate a warning. The warning results in a flaky test failure.

  flush logs;
  purge binary logs before now();
  +Warnings:
  +Warning  1868  file ./0.000002 was not purged because it is the active log file.

Squash with D10120169

Reviewed By: abhinav04sharma

Differential Revision: D10858440

fbshipit-source-id: f502ea9b6f7