Wake workers up before waiting for them to stop
Summary:
While stopping slave workers we should first wake them up and then wait
for them to exit. We used to do this in lock step for each worker which can
cause us to wait on a single worker for ever when that worker depends on a
different worker which has not been woken up yet.

Reviewed By: anirbanr-fb

Differential Revision: D12851975

fbshipit-source-id: 3036bdc4c53