Bug #28825617 ASSERTION FAILURE: LOB0LOB.CC.*TRX == NULLPTR || TRX->IS_READ_UNCOMMITTED()
Summary:
```
Background:

This is a regression caused by the following commit:

commit 56eb89637db93f16bcf7e7e047d6ec7907cd5bbe
Author: Kevin Lewis <kevin.lewis@oracle.com>
Date:   Wed May 23 15:51:08 2018 -0700

Bug#25540277 - INNODB'S MVCC HAS O(N^2) BEHAVIORS

Problem:

The function row_sel_build_prev_vers_for_mysql() goes through the old versions
of the clustered index record and identifies the one which can be viewed by the
current transaction (MVCC).  The result is then cached in an object of
Row_sel_get_clust_rec_for_mysql.

The two things that are cached by the helper class
Row_sel_get_clust_rec_for_mysql are cached_clust_rec and cached_old_vers.  If
the latest record is cached_clust_rec, its old version for the given
transaction that is viewable is cached_old_vers. The important point here is
that the record offsets for these two records are not cached.

The record offsets for cached_clust_rec and the cached_old_vers need not be the
same.  This is confirmed by looking at row_sel_build_prev_vers_for_mysql().
The 6th argument which contains the record offsets is an input/output argument.
The input is the offsets of the latest version of clust_rec, and the output is
the record offsets of the older version.  So the record offsets of different
versions of clust_rec can change.

Solution:

Re-calculate the record offsets for the cached old_vers.

rb#20927 approved by Kevin.
```

Reviewed By: hermanlee

Differential Revision: D13999012

fbshipit-source-id: f41126f0f28