Fix race in killing com_rpc
Summary:
There was a race condition in the kill
sequence where a session might be killed via
`kill ID` but it would still execute queries.
This happens if it is removed from the map while
not attached, but someone already has a reference to it.

Work around this by falling through to the code that sets the kill flag,
so if the race occurs the code that performs the attach can see the KILL
The existing code had an early return to skip the work, but the
work is necessary

Reviewed By: jkedgar

Differential Revision: D13062462

fbshipit-source-id: debb656676c