Disable async client for session tracker test
Summary:
GET_LOCK returns a result set, which will
only have an OK packet if DEPRECATE_EOF is enabled.
This is explicitly disabled for the async client due to
weird interactions. Add an option to disable async client
for this test.

The rest of the queries pass because SET statements etc. already
return OK packets

Squash with: D13541369

Reviewed By: lth

Differential Revision: D13605715

fbshipit-source-id: c3487e02284