FB8-62: Handle Threads_binlog_client with gtid
Summary: the original commit for threads_running in 5.6 contained bugs.
The 8.0 port simply ported these bugs to 8.0.
This commit solves them, as 06c075e did for 5.6.

Instead of following the 5.6 implementation, this commit moves the
variable changes inside the mysql_binlog_send function, ensuring
that counters will always be correctly modified when that function
is called.

Test Plan: modified the mtr test case.