Fix db_metadata not getting reset in db_load_opt if db.opt is of old format
Summary: In the recent sev (s172299) we found out db_metadata was incorrectly overwritten by db_metadata from other tables. The problem is that in my previous change HA_CREATE_INFO wasn't properly cleared and as a result if your db.opt is of old format that does NOT have db_metadata field, such database can incorrectly inherit the db_metadata (or other properties) from previous databases. In the new format case, db_metadata is always there even in the case of empty so this would not happen. This is why my previous testing didn't catch it as it requires specifically creating a db.opt of old format (or empty) and restarting the server to trigger reloading db.opt to repro this issue.

Reviewed By: hermanlee

Differential Revision: D13869337

fbshipit-source-id: 28a2b81e9dc