Fix bulk-load test with non-deterministic snapshot count
Summary:
Rockdb patch (https://github.com/facebook/rocksdb/commit/5d4fddfa52768952e58abae5044174ad0ce30f7d)
now takes a snapshot before compaction starts and release it afterwards for write_prepared feature. Since compaction is a background thread, the snapshot showing up in DB_NUM_SNAPSHOTS becomes non-deterministic.

This test changes logic from asserting on concrete count of snapshot, just start check that there are open snapshots. Note that this is not a critical assert in this test but just a way to verify that the operation above is executed as expected to move the system to some state.

Reviewed By: hermanlee

Differential Revision: D13961245

fbshipit-source-id: b9f516b36df