Fix bug allowing transactions to commit after DB read only
Summary:
During the port, the db_read_only_hash mutex was not held during commit time, and only when checking the hash map.

This could mean we could have a race where a commit thread finishes checking for read only, then ALTER DATABASE sets the DB to read only and returns, then the commit thread continues to commit the transaction. A similar issue exists for DROP DATABASE.

To fix this issue, make use of metadata locking to serialize these operations. The ALTER DB code path now takes an exclusive MDL lock on all tables (similar to the drop code path) while holding the schema lock. This is sufficient to prevent races because `CREATE TABLE` acquires the schema lock as well.

This means that it is now not possible to set DB read only when there are transactions in flight, which is a change of behaviour from before.

There's also a fix for a small bug where `ALTER DB SUPER_READ_ONLY = TRUE` could fail with DB read-only, if its status was not cached in THD.

Squash with: D9379904

Reviewed By: hermanlee

Differential Revision: D14014269

fbshipit-source-id: 74e1a55f864