[mysql-test-run] Add option to skip big tests
Summary:
Our tests are timing out because we're running big tests even when we did not ask for them.

This is because as of https://github.com/mysql/mysql-server/commit/f0bcecb315a2, mtr will now turn on disabled and big tests even if the user did not ask for them. Add an option to override this behaviour.

Differential Revision: D14115272