Fix writing format_description_event with different log_event_types.
Summary:
This fixes https://bugs.mysql.com/bug.php?id=94319

Different versions of mysqld have different number of binlog event
types. The Format_description_event log has an array whose size is based
on the number of events supported by the mysqld version. So a 5.6 mysqld
generates a FDE with 35 types while 8.0 mysqld generates a FDE with 39
types. When 8.0 mysqld attempts to read a 5.6 generated FDE and then
rewrite out the 5.6 FDE, it uses its own value of LOG_EVENT_TYPES (39)
to determine the length of the array, which is actually only 35 types.
The copy accesses undefined memory to generate the FDE.

Fix this by using the length of the FDE recorded in the array instead of
LOG_EVENT_TYPES.

Reviewed By: lth

Differential Revision: D14113482

fbshipit-source-id: 1e5e2b9e59b