FB8-47 : Increase max user name length to 80 (#937) (#937)
Summary:
https://jira.percona.com/browse/FB8-47

The max user name length is increased to 80 chars (utf8mb3)
Since the system tables are InnoDB tables and because Primary key is
on the full username (80 *3 = 240 bytes) and other columns, InnoDb
initialization with 4k page size fails.

This is new limitation.

Reference Patch: https://github.com/facebook/mysql-5.6/commit/3dbbcfc

This mitigates upstream bug: http://bugs.mysql.com/3083

User name length limit of 16 was too short, so made it 80.

80 is effectively the limit without making additional changes
because of MyISAM's 1000 byte limit on key length and the
primary key of mysql.columns_priv. To compute the size of
the key for that table, the lengths of the char fields are
summed and then multipled by 3, resulting in a limit on the
max size of user name lengths being 81 characters.
Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/937

Reviewed By: lloyd

Differential Revision: D13890369

Pulled By: lth