Support symbolizer for clang sanitizer builds
Summary:
```
This is a resubmission of D10108064.

Clang sanitizer builds don't symbolize the stack by default, and their recommended way of doing that is setting ASAN_SYMBOLIZE_PATH env var(https://clang.llvm.org/docs/AddressSanitizer.html). However, setting envvar everytime is tedious so we want to link symbolizer library to the san-binaries so that we can symbolize the stack without the envvar.

We wanted to make the change only from the tools side, but applying it outside of mysqld binary requires adding lots of custom linker flags to pass cmake tests (e.g. manually adding all sanitizer locations along with -lstdc++, -lm, -lz, lpthread, etc to LDFLAGS). However, if the change proposed in this diff is allowed, we can support the symbolizer across all binaries in cleaner fashion (D9975685 is sufficient).

I think this amonunt of change is okay as the enclosing IF() is added by us as well. But I'd like to get your opinion about i as the enclosing IF() is added by us as well. But I'd like to get your opinion about i as the enclosing IF() is added by us as well. But I'd like to get your opinion about i as the enclosing IF() is added by us as well. But I'd like to get your opinion about it.
```

When linking the protoc binary, we need to remove these extra flags, because ubsan is disabled for this binary.

Reference patch: https://github.com/facebook/mysql-5.6/commit/8d12f13f25b68d3d510879c98e0173a835bea146

Reviewed By: hermanlee

Differential Revision: D14271964

fbshipit-source-id: 5a534b47221