Fix ubsan error
Summary:
When table_session_query_attrs_impl is initialized with query attributes, it is possible that there are no attributes in the current query. This means that `m_current_row` gets set to the "first" element of a zero-length vector which does not exist, and causes ubsan to complain. Instead, we should just set this to nullptr, since `m_current_row` will be initialized later anyway during iteration.

Squash with: D9385165

Reviewed By: hermanlee

Differential Revision: D14272341

fbshipit-source-id: 9a9ce239914