Rename CACHE_LINE_SIZE to avoid conflicts with rocksdb. This change aligns with MySQL 8.0
Summary: When testing multi-get I'm noticing CACHE_LINE_SIZE conflict is happening again in a different rdb_*.cc file. We had a workaround in ha_rocksdb.cc to undef it but longer term we should either rename either one in MySQL or RocksDB. It looks like changing MySQL is easier as it is only one file and RocksDB has ~50. So I'm changing MySQL in this diff.

Differential Revision: D14927977 (https://github.com/facebook/mysql-5.6/commit/e0c509bcbe06bdbf22bee1bcdbfef4f9d165e9d7)

fbshipit-source-id: 47588f29639