block_create_no_primary_key ignores tables that don't specify engine
Summary:
When block_create_no_primary_key is enabled, tables without primary keys
should be blocked from being created. However, the check relies on the
engine to the create table statement to be specified. If no engine is
specified, then the create table succeeds.

For the fix, check the default storage when no engine is specified to
see if INNODB or ROCKSDB is being used.

Reviewed By: luqun

Differential Revision: D15047959

fbshipit-source-id: 5158cabcbda