Platform007 test fixes in MySQL 5.6
Summary: So far the only issue I've found are the SSL related test differences. It's not clear to me how much they matter (hence the review). Ignore the test failures for now as they wouldn't work until the tooling has been updated to platform007. It'd be interesting to see how to coordinate the changes.

Reviewed By: hermanlee

Differential Revision: D15060499

fbshipit-source-id: 2af28804965