Fix innodb-mysql
Summary:
It fails occasionally with:

```
[  1%] innodb.innodb_mysql                      [ pass ]  103780
[  2%] innodb.innodb_mysql                      [ pass ]  102875
[  3%] innodb.innodb_mysql                      [ pass ]  108809
[  4%] innodb.innodb_mysql                      [ pass ]  104749
[  5%] innodb.innodb_mysql                      [ pass ]  109692
[  6%] innodb.innodb_mysql                      [ pass ]  102458
[  7%] innodb.innodb_mysql                      [ fail ]
        Test ended at 2019-05-15 22:08:30

CURRENT_TEST: innodb.innodb_mysql
  --- /data/users/yzha/mysql/mysql80-fork/mysql-test/suite/innodb/r/innodb_mysql.result   2019-05-14 21:54:40.757610545 +0300
  +++ /data/users/yzha/mysql/mysql80-fork/_build-8.0-Debug/mysql-test/var/log/innodb_mysql.reject 2019-05-16 08:08:30.406513094 +0300
  @@ -415,7 +415,7 @@
   # Masking (#) number in "rows" column of the following EXPLAIN output, as it may vary (bug#47746).
   EXPLAIN SELECT DISTINCT t1.name, t1.dept FROM t1 WHERE t1.name='rs5';
   id     select_type     table   partitions      type    possible_keys   key     key_len ref     rows    filtered        Extra
  -1      SIMPLE  t1      NULL    ref     name    name    82      const   #       100.00  Using where; Using index
  +1      SIMPLE  t1      NULL    range   name    name    164     NULL    #       100.00  Using where; Using index for group-by
   Warnings:
   Note   1003    /* select#1 */ select distinct `test`.`t1`.`name` AS `name`,`test`.`t1`.`dept` AS `dept` from `test`.`t1` where (`test`.`t1`.`name` = 'rs5')
   SELECT DISTINCT t1.name, t1.dept FROM t1 WHERE t1.name='rs5';

mysqltest: Result content mismatch
```

My guess is that innodb uses sampling stats so it might end up choosing a different query plan and even though the table should be empty, the stats didn't get updated in time. Adding ANALYZE TABLE seem to have fixed it. Before the fix I can repro it with < 10 tries and after the fix it ran for 250+ times in total and still haven't failed yet.

Looking at the original bug and its fix:

https://lists.mysql.com/commits/19658
https://bugs.mysql.com/bug.php?id=26159

It isn't really an optimizer bug so the EXPLAIN SELECT statement there doesn't seem to be particularly useful. Having said that, I'm keeping it there for now (just in case I was wrong and in order to preserve the original intent as much as possible) and will only remove it as a last resort.

Reviewed By: hermanlee

Differential Revision: D15374637

fbshipit-source-id: 0ec0bb5573c