Improve statistics collection in MyRocks
Summary:
Currently index stats calculation is based on ssts and memtables. It has a number of issues as detailed in T27908367. The change aims to address these issues. The general approach is to do index stats calculation at MyRocks level, rather than calling into RocksDB to collect index stats of sst files and memtable. This way we can resolve the issues that involve sst file and memtable.

At MyRocks level, we can calculate stats for an index by a full scan using rocksdb::Iterator. There are 2 ways to trigger an index scan: 1. analyze table and 2. background thread. This change creates a new background thread for stats calculation. For each table, when the number of rows modified reaches a threshold, for example, 10% of total rows, this table is inserted into a queue of the background thread. The thread will pick up the table and do index scan. This approach is similar to InnoDB's index calculation design.

The cost of index scan is higher than the current approach. To address that we do sampling and also allows interrupting the scan at per-key level. For example when the server shuts down or the analyze table thread receives a kill signal.

This change refactor ha_rocksdb::info(). Currently the method calls GetApproximateMemTableStats(), which is expensive. This change removes the call and let the background thread to call it instead.

This change adds a system variable for switching between the old and new index calculation approach.

Differential Revision: D15406669

fbshipit-source-id: c2200982cd0