Fix innodb_stats_on_metadata corrupting index stats on partitioned tables
Summary:
The innodb_stats_on_metadata is not correct for partitioned tables.

For partitioned tables, the `info(HA_STATUS_CONST)` call is only done on the largest partition while for most other flags including `HA_STATUS_VARIABLE`, the call is done for all partitions in a loop.

The issue is that `innodb_stats_on_metadata` implicitly converts `info(HA_STATUS_VARIABLE)` calls into `info(HA_STATUS_VARIABLE | HA_STATUS_CONST)` calls. Since this is done in a loop, it means that we're calculating index stats for just the last partition, giving inaccurate statistics.

There's no good way to fix it, because we either need InnoDB to understand partitions, or ther other way around. The latter approach was chosen, so `innodb_stats_on_metadata` is made a server session variable, and ha_partition will do the correct thing if it detects the variable is on, and InnoDB tables are used.

Because MySQL 8 has native partitions, this is a not an issue there, as the native partition code has access to both the innodb_stats_on_metadata variable, and the algorithm for processing info for partitioned tables.

Reviewed By: luqun

Differential Revision: D15632110

fbshipit-source-id: cfdeefcf9ae