Extending SHUTDOWN query to support read_only/aborting
Summary:
This diff extends SHUTDOWN query to support the following
features.
- Aborting with any specified exit code (range is 0..255).
If nothing is specified or 0 is given, it does default clean
shutdown. If 1+ is given, exits with the given error code
immediately. This is helpful to shutting down instance
even if it is stuck somewhere.
- Aborting on read_only instances only (typically replica instances).
If read_only option is given, and if the instance is read_only==0,
the command fails before executing shutdown.

Reviewed By: luqun

Differential Revision: D15749084

fbshipit-source-id: 5354b969fa6