Fix valgrind errors on rocksdb.skip_core_dump_on_error
Summary:
A valgrind error was happening in xid_to_str because it was trying to read the whole xid->data buffer, but we don't populate the whole buffer, leaving bytes at the end uninitialized. Fix by zero'ing the buffer.

This test is also not suitable for valgrind because it does crash testing, causing many memory leaks. Disable the test for valgrind.

Squash with: D15395550

Reviewed By: zhichengzhu

Differential Revision: D15806281

fbshipit-source-id: cec5fe95611