Fix dangling SLAVE_INFO/THD in slave_list in case of slave communication failure
Summary:
The global slave_list in master is inserted when a slave calls REGISTER_SLAVE on master, and removed when COM_BINLOG_DUMP_GTID is finished execution. If for some reason, the slave fails to communicate with the master, it is possible the slave_list didn't get a chance to be updated, and therefore leaking the SLAVE_INFO with a dangling THD pointer (that gets deleted earlier when the connection disconnects). If the slave later reconnects with the same server_id, it could ovewrite the same entry by unregister and registering again, and end up fixing this leak. However, this is still problematic because:
1. You'd have a "brief" window where slave_list contains the leaked SLAVE_INFO with dangling THD,
2. If the slave connects with a different server_id or never reconnect, you'd stuck with the leaked SLAVE_INFO and dangling THD pointer.

In either case, socket_diag_slaves or show slave hosts could access this dangling THD pointer and crash. In the actual crash dump, the wormhole is the "slave" and it always reconnects with a new random server_id, which is effectively #2.

Given that the slave may never reconnect again, the right fix is to have the THD unregister itself only if the THD in the list with the same server_id matches this particular THD (so that it won't race with the unregister/register in case the slave reconnects with a new THD).

The unregistration can happen after the cleanup during LOCK_thd_data. This way, the fill_socket_diag_slaves code would either see the empty VIO and stop proceeding, or seeing the slave_list data before the THD data gets released. The slave_list could still be "dangling" and the data could be stale, but it is "valid" enough that the only pointer are THD and rest are character arrays/integer/etc, so fill_socket_diag_slave can proceed safely.

NOTE: I've avoided changing SHOW SLAVE HOSTS for now and it needs to have the same set of locking l(locking the global THD lock, individual THD_data lock, etc) like fill_socket_diag_slaves do given that it touches THD as well. But in order to lower the risk of this change I'm avoiding SHOW SLAVE HOSTS command for now and will address it separately.

Reviewed By: lth

Differential Revision: D15860487

fbshipit-source-id: ae15f9940bf