Repurpose SLAVE_EVENT as METADATA_EVENT
Summary:
Repurpse SLAVE_EVENT as METADATA_EVENT. This new METADATA_EVENT will follow
TLV encoding - this will ensure that it is easy to extend this event in future.

This new method has a 64 bit flags in the data-header part (header is not
encoded as TLV). The data-body currently consists of a 64 bit hlc timestamp.

This needs to be ported to 8.0 before using in prod since we use 8.0 mysqlbinlog/mysqlclient in prod

Reviewed By: abhinav04sharma

Differential Revision: D15733469

fbshipit-source-id: 0aa5053a4d9