Properly support SSL options in mysqlslap
Summary: Mysqlslap doesn't apply options the same way for the initial connection and the workers - as a result the workers missing a lot of options passed from command line, including SSL. This fixes the SSL part only. I'll address the bigger issue with probably a bit refactoring with a separate fix (which is going to require quite a bit more testing).

Differential Revision: D16015149