Assign HLC timestamp during ordered commit and write to binlog
Summary:
Adds the ability to assign HLC timestamp and write to binlog. HLC is written as
part of Metadata_log_event which comes immediately after Gtid_log_event in
binlog.
1. On a master instance, the Metadata_log_event is written to binlog during
the begin of a trx and after (dummy) GTID event is written. This is just a
placeholder to reserve space in cache (similar to how a dummy GTID event is
written in cache during the begin of the txn)
2. On a master instance, the real commit time HLC timestamp is written during
ordered commit's binlog flush stage. The HLC timestamp is generated after
assigning a real GTID for the trx. This ensures that gtid1 > gtid2 <==>
hlc_gtid1 > hlc_gtid2;
3. On a slave instance, the apply of a Metadata_log_event just stashes the
HLC timestamp generated in the master in THD variable. When the trx is actually
applied we extract the HLC timestamp from THD variable and write it to binlog
i.e a slave retains the HLC timestamp associated with the trx (as generated by
the master)
4. On a slave instance, after engine commit, we bump the local HLC clock to
be greater than the HLC commit timestamp of this trx (i.e the commit timestamp
of the trx as generated by the master). This ensures that if this slave becomes
a master in future, then it starts assigning HLC timestamps which is greater
than any previous trx's HLC timestamp i.e monotonocity of trx's HLC timestamp is
preserved across promotions)

Persistence/crash recovery of HLC values will come in subsequent diff

Reviewed By: abhinav04sharma

Differential Revision: D15733493

fbshipit-source-id: 50a6c059f21