Integrate raft plugin with ordered_commit in mysql
Summary:
During ordered commit, we have to invoke the before_flush hook to plugin.
This should happen after GTID is updated, but before writing/flushing the THD's
IO_CACHE to binlog. The diff adds the support to do the same in the flush stage
of ordered commit. Note that, to begin with we will continue to write/flush to
binlog in mysql server (after plugin returns). This will eventually be removed
and ordered_commit will be refactored accordingly.

Reviewed By: anirbanr-fb

Differential Revision: D15866650

fbshipit-source-id: 17cdc0a5ec6