Disable mysqlx plugin in tests by default
Summary:
This is a fun one.

We've been observing random unexplained connections from the test suite, something like the following:

```
Audit_null_connection_change_user	0
-Audit_null_connection_connect	10
-Audit_null_connection_disconnect	10
-Audit_null_connection_pre_authenticate	10
+Audit_null_connection_connect	11
+Audit_null_connection_disconnect	11
+Audit_null_connection_pre_authenticate	11
```

There is a rogue connection somewhere into the process. However we weren't able to find out exactly where it came from.

Then we lucked out: in one test case it is failing and print out parts of the query:

```
@@ -25,6 +25,7 @@
 WHERE line LIKE ( '%SELECT%' );
 incoming_query
 'The_original_query'
+@tls_version
 DROP TABLE test_log;
 SET GLOBAL general_log = DEFAULT;
 SET GLOBAL general_log_file=  old_general_log_file;
mysqltest: Result length mismatch
```

The extra tls_version is the last section (separated by ' ') from the following SELECT in the test:

```
  SELECT substring_index( line, ' ', -1 ) AS incoming_query
  FROM test_log
  WHERE line LIKE ( '%SELECT%' );
```

This means we are looking for a SELECT with tls_version query, and it is coming from mysqlx:

```
      sql_result.query(
          "SELECT skip_networking, skip_name_resolve, have_ssl='YES', "
          "ssl_key, "
          "ssl_ca, ssl_capath, ssl_cert, ssl_cipher, ssl_crl, "
          "ssl_crlpath, tls_version;");
```

The query itself is coming from mysqlx startup and it is done from another thread asynchronously, which explains the race condition where we would only see it randomly popping up in test cases. This also explains why we only see it in 8.0 but never in 5.6.

For now, the easy thing to do is to disable mysqlx globally in the test suite, and fortunately the mysqlx plugin test suite has a x/t/suite.opt that enables mysqlx by default and that overrides the --skip-mysqlx.

I've verified that mysqlx is disabled in other test cases (by inserting a SHOW PLUGINS)

```
+mysqlx DISABLED        DAEMON  NULL    GPL
+mysqlx_cache_cleaner   ACTIVE  AUDIT   NULL    GPL
```

and enabled properly in mysqlx test cases (by inserting a SHOW PLUGINS)

```
+mysqlx ACTIVE  DAEMON  NULL    GPL
+mysqlx_cache_cleaner   ACTIVE  AUDIT   NULL    GPL
```

Note that I also had to do a few things besides globally disable mysqlx:
1. enable mysqlx for sys_var suites because they test mysqlx variables and those tests are safe as they don't do any show processlist stuff
2. Rebaseline tests that has mysqlx artifacts. For reference, the original commit that enabled mysqlx across the board is this commit: aed56f7708413ffe5ee600a20bea838875173440, WL#11538.
3. make xplugin_wait_for_interface.inc more resilient by doing nothing if xplugin is not enabled

Keep in mind this will lose some X plugin coverage in non-X-plugin scenarios (xplugin still has its own test suite) but I doubt it matters that much for us.

Differential Revision: D16106437 (https://github.com/facebook/mysql-5.6/commit/57c33122ba5c461e488609fd25835e32044cdd4a)

fbshipit-source-id: fbc3a9ae4d4