Return hlc timestamp for a write txn
Summary:
The HLC timestamp associated with a GTID of a txn is returned as part of OK
packets. The HLC timestamp is stored in session response attributes tracker
field of the OK packet. The HLC is stored as a key-value pair in session
response attribute - the key is 'hlc_ts' and the value is the HLC timestamp.
Setting the HLC timestamp in OK packets is gated by a sysvar
'session_track_hlc'.

Only a RW transaction (i.e a transaction that genarates a GTID) will have a HLC
timestamp retuned in its OK packet.

Reviewed By: jrahman

Differential Revision: D15987483

fbshipit-source-id: 570dcfd0af9