Commit time hooks support in mysql
Summary:
For synchronization between raft plugin and mysql to work correctly, mysql has to
know the transaction as represented in the raft world. So the OpID (term, index)
tuple need to be propagated back to mysql from raft after a replication round is
initiated i.e in the before_flush hook of the Binlog storage observer. This
(term, index) will be sent back to raft plugin in the before_commit hook of
Transaction observer (which would be stored in the THD after before_flush hook
returns)

This diff adds the support for the same. replication.h will be pulled into
fbcode once this diff lands.

Reviewed By: abhinav04sharma, anirbanr-fb

Differential Revision: D16232239

fbshipit-source-id: 99086bc620d