Use std::unique_ptr instead of shared_ptr for Rdb_converter
Summary:
Originally, std::shared_ptr was used because for std::unique_ptr, the destructor needed full type definition of Rdb_converter. For std::shared_ptr, the full type was required at construction instead of destruction, so we used that as a workaround.

A better fix would have been to just move ~ha_rocksdb to the .cc when full type definition is available, so that it knows how to cleanup the Rdb_converter object.

Reviewed By: luqun

Differential Revision: D16502955 (https://github.com/facebook/mysql-5.6/commit/7e67c47d6e227cecc2cfabd90a2f1f1890d8a630)

fbshipit-source-id: fb52095ec22